Move integrate_widget into pyFAI.gui
